requirements.txt
Flask==2.3.3
pyvmomi==8.0.2.0.1
Flask-WTF==1.1.1
python-dotenv==1.0.0
Flask-Login==0.6.2
pandas==2.1.1
plotly==5.17.0
requests==2.31.0
pywin32==306; sys_platform == 'win32'
waitress==2.1.2
